Renamed progress_observer to progress_delegate.py Shifted ThreadedProgressObserber to xcube.util.progress
