+++ xcube/cli/_gen2/progress_delegate.py
@@ -18,21 +18,13 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-import threading
-from time import sleep
-from timeit import default_timer
-from typing import Sequence, Any
+from typing import Sequence
 
 import requests
 
 from xcube.cli._gen2.genconfig import CallbackConfig
 from xcube.util.assertions import assert_condition, assert_given
-from xcube.util.progress import ProgressObserver
 from xcube.util.progress import ProgressState
-
-
-# Helge, please have a look at impl. dask.diagnostics.progress.ProgressBar
-# It uses a separate Thread() to only update a progress bar within fixed time deltas.
 
 
 def _format_time(t):
@@ -52,87 +44,6 @@ def _format_time(t):
         return "{0:2.0f}min {1:4.1f}s".format(m, s)
     else:
         return "{0:4.1f}s".format(s)
-
-
-class ThreadedProgressObserver(ProgressObserver):
-    """
-    A threaded Progress observer adapted from Dask's ProgressBar class.
-    """
-    def __init__(self,
-                 delegate: Any,
-                 minimum: float = 0,
-                 dt: float = 1):
-        """
-
-        :type dt: float
-        :type minimum: float
-        """
-        super().__init__()
-        assert_condition(dt >= 0, "The timer's time step must be >=0")
-        assert_condition(minimum >= 0, "The timer's minimum must be >=0")
-        self._delegate = delegate
-        self._running = False
-        self._start_time = None
-        self._minimum = minimum
-        self._dt = dt
-        self._width = 100
-        self._current_sender = None
-        self._state_stack: [ProgressState] = []
-
-    def _timer_func(self):
-        """Background thread for updating the progress bar"""
-        while self._running:
-            elapsed = default_timer() - self._start_time
-            if elapsed > self._minimum:
-                self._update_state(elapsed)
-            sleep(self._dt)
-
-    def _start_timer(self):
-        self._width = self._state_stack[0].total_work
-        self._start_time = default_timer()
-        # Start background thread
-        self._running = True
-        self._timer = threading.Thread(target=self._timer_func)
-        self._timer.daemon = True
-        self._timer.start()
-
-    def _stop_timer(self, errored):
-        self._running = False
-        self._timer.join()
-        elapsed = default_timer() - self._start_time
-        self.last_duration = elapsed
-        if elapsed < self._minimum:
-            return
-        if not errored:
-            self._delegate(self._current_sender, elapsed, self._state_stack)
-        else:
-            self._update_state(elapsed)
-
-    def _update_state(self, elapsed):
-        if not self._state_stack:
-            self._delegate(self._current_sender, 0, elapsed)
-            return
-        state = self._state_stack[0]
-
-        if state.completed_work < state.total_work:
-            self._delegate(self._current_sender, elapsed, self._state_stack)
-
-    def on_begin(self, state_stack: Sequence[ProgressState]):
-        assert_given(state_stack, name='state_stack')
-        self._state_stack = state_stack
-        self._current_sender = "on_begin"
-        self._start_timer()
-
-    def on_update(self, state_stack: Sequence[ProgressState]):
-        assert_given(state_stack, name="state_stack")
-        self._state_stack = state_stack
-        self._current_sender = "on_update"
-
-    def on_end(self, state_stack: Sequence[ProgressState]):
-        assert_given(state_stack, name="state_stack")
-        self._state_stack = state_stack
-        self._current_sender = "on_end"
-        self._stop_timer(False)
 
 
 def get_callback_api_progress_delegate(callback_config: CallbackConfig):
